Added LinkedInService class instead of LinkedIn Integration
