+++ SocialIntegration.xcworkspace/xcuserdata/neosoft1.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -190,8 +190,8 @@
             filePath = "SocialIntegration/Controllers/ViewController.swift"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "102"
-            endingLineNumber = "102"
+            startingLineNumber = "107"
+            endingLineNumber = "107"
             landmarkName = "clickedTwitterLoginButton(_:)"
             landmarkType = "7">
          </BreakpointContent>
